Touchstone auth blueprint and SAML flow
Why these changes are being introduced:

This project is designed to expose a Flask blueprint and helper utilities
that a different Flask app can install and import to require authentication
for selected Flask routes.

While this repository also contains a standalone Flask app for debugging
and development, it is the library installed from touchstone_auth_flask/lib
that is the primary product.

How this addresses that need:
* Defines an auth oriented Flask blueprint 'touchstone_auth_bp'
* Flask-Login utilized to store logged in users in Flask session
* SAML integration with Touchstone established under /saml routes in blueprint
* root level setup.py created that allows pip install of this project, resulting
in a 'touchstone-auth-flask` library

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1008