Set some IdP and SP configurations dynamically
Why these changes are being introduced:

Information like the IdP x509 certificate and SSO URL can be retrieved
from the IdP metadata.  Retrieving this dynamically makes it more flexible to
IdP changes, and reduces the need to store sensitive information even in AWS
params.

We must also provide the SP certificate and key.  For local development
formerly this required generating a self-signed certificate and setting that as
environment variables.  It would be handy if when locally developing
this was handled automatically.

How this addresses that need:
* Retrieves IdP SSO URL and x509 certificate once and saves to Flask app config
* Dynamically creates self-signed certificate for local development for SP

Side effects of this change:
* IdP SSO URL and certificate will not be required as AWS parameters

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1008